# perf: x/tx/signing/textual: remove unnecessary,expensive regex from method