--- a/x/tx/signing/textual/message.go
+++ b/x/tx/signing/textual/message.go
@@ -261,10 +261,8 @@ func (mr *messageValueRenderer) Parse(ctx context.Context, screens []Screen) (pr
 	return protoreflect.ValueOfMessage(msg), nil
 }
 
-var (
-	// <int> <field_kind>
-	headerRegex = regexp.MustCompile(`(\d+) .+`)
-)
+// <int> <field_kind>
+var headerRegex = regexp.MustCompile(`(\d+) .+`)
 
 func (mr *messageValueRenderer) parseRepeated(ctx context.Context, screens []Screen, l protoreflect.List, vr ValueRenderer) error {
 	res := headerRegex.FindAllStringSubmatch(screens[0].Content, -1)