This change hoists expensive regexp.Regexp objects from (*message*Renderer).Parse where they were created by constant strings, moreover in a well used method, and instead makes them globals that are initialized exactly once then invoked whenever needed. Noticed in a fuzzing and auditing session, where they stood out trivially and are our rule book at Orijtech Inc for common simple things to watch out.

The reduction in CPU time, allocations is drastic as shown below:

```
$ benchstat before.txt after.txt
name                         old time/op    new time/op    delta
MessageValueRendererParse-8    76.0µs ± 3%    61.0µs ± 1%  -19.75%  (p=0.000 n=10+10)

name                         old alloc/op   new alloc/op   delta
MessageValueRendererParse-8    55.1kB ± 0%    38.9kB ± 0%  -29.38%  (p=0.000 n=10+10)

name                         old allocs/op  new allocs/op  delta
MessageValueRendererParse-8       696 ± 0%       564 ± 0%  -18.97%  (p=0.000 n=10+10)
```

Fixes #15765